Get resolved root entity from parent main-meta
The core library now stores the resolved root entity in the parent
main-meta rather than in the root-meta, for the following reason:
main composes the root entity like a field composes an entity.
So the resolved root entity is stored in the Resolved aux of mainMeta
like a resolved entity is stored in the Resolved aux of its parent
fieldMeta. This makes things easier and consistent.