Add foreign-keys from immediate children to root entity
Previously, such foreign-keys were added only in keyless entities. Now
it is added to all immediate children, even if they have keys. This
ensures that the root entity cannot be deleted if it has any children,
irrespective of whether these children have keys or not.

The flip side is that the root entity must be created before children
can be created (they can be in the same create-request though).