Drop runtime info from aux plugins
Aux plugin classes need to be specified in tree-ware configuration in
`build.gradle.kts` so that they can be included in the code generated
by the core tree-ware Gradle plugin. It is hard for the generated code
to refer to runtime information. So the `environment` parameter has
been dropped from `MySqlMetaModelMapAuxPlugin`. But the environment
name is still needed in case it needs to be used as a prefix for the
database name. So a `databasePrefix` parameter has been added to all
the places that deal with the database name.